// project2/myapp/src/model/student.tsx
export default interface ISudent {
  id: number;
  nameSchool: string;
  district: string;
  codeStudent: string;
  classroom: string;
  fullname: string;
  date: number;
  month: number;
  year: number;
  gender: boolean;
  birthplace: string;
  ethnic: string;
  permanentResidence: string;
  phoneNumber: string;
  point1: number;
  point2: number;
  point3: number;
  point4: number;
  point5: number;
  sumCore: number;
  priorityPoint: number;
  totalPoint: number;
  note: string;
}